ddns: extract pkg/ddns spine from pkg/dhcpserver (#2691 P1a)
Phase P1a of the #2691 world-class DDNS redesign: a verbatim,
no-behavior-change code move of the Dynamic DNS spine out of
pkg/dhcpserver into a new pkg/ddns package, so Surface A (router/
interface-address publish) and Surface B (DHCP-lease publish) can later
share one engine/backend/state without smearing DHCP DHCID/lease
semantics onto router records (plan §7 Fork 1, recommendation (b)). This
is the foundation P1b-P3 build on; it changes no behavior — every moved
test keeps its assertions intact.

What moved to pkg/ddns (logic + tests verbatim):
  - the reconcile engine (DDNSManager -> ddns.Manager): policyFromConfig,
    Reconcile, reconcileOnceLocked, upsertLocked/deleteOwnedLocked,
    withdrawAllLocked, ownerWatermark, Stats, OwnedRecordView(s), the
    write-ahead durability + never-delete-non-owned boundary
    (ddns.go -> manager.go);
  - the ownership state store (ddns_state.go -> state.go);
  - the DNSUpdater interface + LeaseDNSRecord + record/PTR-name helpers
    (ddns_dns.go -> backend.go);
  - the LIVE RFC 2136 backend incl. the errDDNSConflictRefused /
    errDDNSPTRPending sentinels, DHCID/replace-owned, TSIG
    (ddns_rfc2136.go -> backend_rfc2136.go);
  - hostname normalization (ddns_hostname.go -> hostname.go);
  - the spine tests (ddns_test.go spine portion -> manager_test.go,
    ddns_rfc2136_test.go, ddns_durability_test.go,
    ddns_manager_inc2_test.go).

What stays in pkg/dhcpserver (glue / DHCP-server-specific):
  - the Kea-memfile lease parser (ddns_leases.go: parseActiveLeases4/6,
    identity4/6, keaLeaseType* constants) — it is entangled with the
    lease-sync memfile fallback (lease_sync.go, #2239/#2262), which is
    DHCP-server-specific, not DDNS-specific (plan §5.10);
  - its tests (ddns_test.go parser portion -> ddns_leases_test.go);
  - a thin glue ddns.go: re-exports the cross-package type aliases
    (DDNSManager, DNSUpdater, LeaseDNSRecord, DDNSStats,
    DDNSOwnedRecordView) so pkg/daemon/grpcapi/cli/api compile unchanged,
    and injects the parser into the engine via the new ddns.LeaseParser
    seam.

Import-cycle handling: the dependency is inverted — pkg/ddns owns the
engine + the Lease record + the LeaseParser seam; pkg/dhcpserver supplies
the Kea parser (keaLeaseParser, projecting ddnsLease onto ddns.Lease) at
construction. pkg/ddns never imports pkg/dhcpserver (verified via
go list -deps); it imports only pkg/config and pkg/fsatomic.

Exports widened only as the package boundary required: DDNSManager ->
ddns.Manager, DDNSStats -> ddns.Stats, DDNSOwnedRecordView ->
ddns.OwnedRecordView, NewManager/NewProductionManager/
NewManagerForTesting, the Lease record + LeaseParser type, and two
test-only accessors (OwnedForTesting/OwnedKeysForTesting) the
cross-package real-parser integration tests use in place of the
now-unexported state store. The HA writer gate (ddnsWriterGateOpen,
pkg/daemon) is unchanged and unmoved — the move touches it only as a
call-through.

Validation: go build ./... clean; go vet ./pkg/ddns/...
./pkg/dhcpserver/... clean; go test (incl. -race) on ./pkg/ddns/...
./pkg/dhcpserver/... ./pkg/daemon/... ./pkg/grpcapi/... ./pkg/cli/...
./pkg/api/... all green; gofmt clean; git rename-detection shows the
move. make test-failover deferred to the parent (no cluster access in
this environment); the gate change is call-through-only.

Refs #2691.
