+++ pkg/ddns/backend_rfc2136_test.go
@@ -1,4 +1,4 @@
-package dhcpserver
+package ddns
 
 import (
 	"context"
